# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CarRentale",
        "version": "1.0"
    },
    "paths": {
...
        "/api/payments": {
            "get": {
                "tags": [
                    "payments"
                ],
                "summary": "Display a listing of payments",
                "operationId": "826dc6b24dd704607aa018a59207dd96",
                "responses": {
                    "200": {
                        "description": "Display a listing of payments"
                    }
                }
            },
            "post": {
                "tags": [
                    "payments"
                ],
                "summary": "Store a newly created payment",
                "operationId": "5e2794ba3d3c64e70873367505e4a751",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "amount",
                                    "rental_id",
                                    "status"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "rental_id": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment created successfully"
                    },
                    "500": {
                        "description": "Stripe API error"
                    }
                }
            }
        },
        "/api/payments/{id}": {
            "get": {
                "tags": [
                    "payments"
                ],
                "summary": "Display the specified payment",
                "operationId": "a9a7656e0f22931dbcedc57a2d959554",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display the specified payment"
                    }
                }
            },
            "put": {
                "tags": [
                    "payments"
                ],
                "summary": "Update the specified payment",
                "operationId": "9f607012ea3accda178adb489aa28da2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "rental_id": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "payments"
                ],
                "summary": "Remove the specified payment",
                "operationId": "eee625edf09d69ad0d35b893a3b4f087",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment deleted successfully"
                    }
                }
            }
        },
        "/api/rentals": {
            "get": {
                "tags": [
                    "rentals"
                ],
                "summary": "Display a listing of rentals",
                "operationId": "a19748eac97c416a4a02a73fbcdb4700",
                "responses": {
                    "200": {
                        "description": "Display a listing of rentals"
                    }
                }
            },
            "post": {
                "tags": [
                    "rentals"
                ],
                "summary": "Store a newly created rental",
                "operationId": "7e700bb6638cc9e6a42aa112bf018ba6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "car_id",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "car_id": {
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "type": "string"
                                    },
                                    "end_date": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rental created successfully"
                    }
                }
            }
        },
        "/api/rentals/{id}": {
            "get": {
                "tags": [
                    "rentals"
                ],
                "summary": "Display the specified rental",
                "operationId": "7248fde8ecc49afa506485b780b4cd27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display the specified rental"
                    }
                }
            },
            "put": {
                "tags": [
                    "rentals"
                ],
                "summary": "Update the specified rental",
                "operationId": "45134c777261098392b52d9f89a0311a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "car_id",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "car_id": {
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "type": "string"
                                    },
                                    "end_date": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rental updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "rentals"
                ],
                "summary": "Remove the specified rental",
                "operationId": "f130db03982a2f6048b833c6cc977fcf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rental deleted successfully"
                    }
                }
            }
        }
    },
...
}